app/django_service/auth_service/requirements.txt
asgiref==3.6.0
protobuf==4.21.6
Django==3.2.16
grpcio==1.49.0
grpcio-tools==1.49.0
djangorestframework==3.13.1
djangogrpcframework==0.2.1
six==1.16.0
sqlparse==0.4.4
psycopg2-binary==2.9.7
gunicorn==20.1.0
PyJWT==2.9.0
requests==2.32.3